Sensors/skeleton: First draft of the skeleton for sensor data collection.
At this point we're still defining the architecture of the project. Mainly how we're going to comunicate between the "blocs" (or nodes if we decide to use microROS and ROS2)

These files shouldn't change much since either way, there should be a wrapper for communication, but depending on how it's done and unforseen implementation choices the types or public methods could change a bit.