Modified platformio.ini for the project to work and removing the old main
The platformio.ini file was not working as it was before.  We had to make changes like adding an upload_port (you may have to change this depending on your own ports). The test section is also added to run separatly the test or the main (this should not be changed).