Updates
Test on gyro is good ! Accurate datas.
For the acceleration on x we don't have correct data yet. The problem seems to be that when we compute the "true" acceleration (ithout the bias) we don't take into account that there is potentially an acceleration on other axes that false the data. We should use quaternion position to make it more accurate (to be done).
Not tested yet for y but it will probably be the same error.