pure_pursuit.cpp: find_lookahead_point (#43)
Adding lookahead and path intersection, instead of taking the first waypoint out of the lookahead distance... In other words, implementing the actual pure-pursuit algorithm instead of simplifying it.
Intuitively I thought it wouldn't affect much, but it actually does, since we're divising by `Ld_fixed_` and not by the radius of the point's destination.