almost working version
Pb:
- lidar crashes after a certain time, possibly because of a maneuver
- the values of the global path overflow on y about a certain time too.
- the connection to the tcp is way too long from my computer, sometimes i can't even see the map before it stops working. Maybe this is just an issue from my computer.

The display of the map is good, the resolution is of 0.4 on a map of 25*25 so it maps a zone of 10m per 10m.

The goal is well updated depending on the frontiers.

The global is also updating correctly, generating a new path when the previous is reached, or the map has been discovered around the goal.